First step in merging code for all 4 DSPs into one single project.
This is utilizing "#defines" to split up the code for the different DSPs. This will hopefully assist in the ability to track changes for all the DSPs and
allow as much reuse of code as possible. This first step defines rk1b2b
and rk2b2b referring to racks 1 and 2 back-to-back 2-level converters
(AFE+INV=back-to-back=b2b). One define should be left uncommented at the
top of main.c to select which code will be compiled. Later in the code
there will be a "#ifdef" rk1b2b then an "#endif", all that code will be
compiled if rk1b2b was defined at the top of the file. If not the code
will be ignored during compilation. Further commits will work on possible
differences in the two converters such as CANbus message IDs, input
inductor value differences, etc.

Eventually the NPC code will also be
integrated into this project with rk1npc and rk2npc defines and the
corresponding ISRs.