Changed back to single timer_isr function for rack 1 and 2 b2b.
Had previously made a duplicate of the entire timer_isr function with
"#if rk1b2b" if statements for each one selecting which rack you wish to
compile to. This reduces the reuse of code so it may be better to only
have one timer_isr for the back-to-back converters and split up any
differences with local "#if"s so as not to have to make a change twice if
both racks need the change.