Standardized IoCU Rm resistors for Iabc of AFE,INV
Previous design for 1kVA per rack had different Rm resistors in the ADC
signal conditioning circuits for AFE and INV in order to best utilize the
measurable range of the ADC. This allowed a higher current possible to be
measured by the AFE channels.

Higher current needs and the desire to standardize some of the signal
conditioning gains, circuits, etc. led to a change today in the IoCU Rm
resistors for racks 1 and 2 AFE/INV. The Analog Input Mapping.xlsx was
also updated to the Sharepoint site.

Now all ADC channels on the IoCU board have Rm = 357||357 = 178.5 ohm
except for channels B7,A5,A7 which have Rm open to measure vabc references
from RTDS, respectively.

This modifies the iabc current measurement scaling in the DSP code. See
accompanying Powerpoint slides for the NLDL tutorial series for notes on
how these scale factors are calculated. Some comments in the code have
been updated to reflect this change as well.

R1 on the voltage transducer boards are also standardized across the AFE
and INV now with 24.75 kohm for va,vb,vc measurements and 39 kohm for the
one DC voltage measurement on the AFE.

Some old comments in the code were also removed.