First commit in ONRdemo branch.
This branch is meant to quickly bring the racks to a state where
all 4 DSPs can be programmed from the same CCS project. This means
having NPC code (PWMs disabled right now) to send the DAC values to
dSPACE as well as individual PWM enables for AFE and INV separated
for racks 1 and 2. This can be seen by the additional #ifdef npc
and #ifdef b2b.