Modified ePWM3 for 240 deg phase shift.
Made the changes to ePWM3 as prescribed by the TI
ePWM user manual for 3phase interleaved DC-DC
http://www.ti.com/lit/ug/sprug04a/sprug04a.pdf
section 3.8

Testing on Saleae Logic16:
ePWM3A shows a 66.68us delay from the ePWM1A rising
edge.

HOWEVER! Just remembered that this DSP still has to
perform the duty of controlling the AFE and the
interleaved DC-DC converter is performed by
repurposing of the INV - this changes how the ePWM
channels are mapped. PWM channels 1-3 should stay the
same as they are used by the AFE whereas channels
4-6 should perform the 3phase interleaved DC-DC task.

Next commit will make the reversion of ePWM1-3 back
to the way they were and make ePWM4-6 have the
settings just tested in this commit and the previous one.