Added functionality of Bormann's Lab 2 Module 9
Change a few things to Bormann's Lab 2 from Module 9
since I'm not using the CPU timer.

I moved the first transmit prior to the while(1) loop
and also added an else statement to the ISR so it would
reset the index back to 0 to repeatedly transmit the message.

This was tested with Putty successfully.