Reading numbers from Putty
Still needs work but here's the stopping point for tonight.

Basically need to decode an array of chars into
a usable number for using as a frequency value.

Might be an error register that gets set when
I closed Putty because the debugger stops
reading in new values when I try to close
Putty and use Matlab instead.