Added notch filter - unsuccessful
Added a notch filter at 929 Hz to the voltage reference
to try to mitigate the resonant frequency of the LC
filter but similar results: the open-loop voltage reference
is being amplified by the LC filter at the resonant frequency.

Next step is to try either a different filter or add on
a PIR controller to get the output voltage I want
across the whole frequency range I want to excite over.